feat(policy): add checks for header case and last character
It's common for commit message guidelines to include restrictions on the
leading case and trailing punctuation of the header; this adds checks to
enforce those. The desired case can be specified as either "upper" or
"lower"; the punctuation check is implemented as a general check for
disallowing the last character in the header to be from any given set of
characters.

Now that there are several header-related checks, this also adds a level
to the config to put all of those checks in one place in the
configuration.

Signed-off-by: Danny Zhu <dzhu@dzhu.us>